+++ src/frontend/src/tests/lib/utils/tokens.utils.spec.ts
@@ -4,7 +4,7 @@ import {
 	ICP_NETWORK_ID
 } from '$env/networks/networks.env';
 import { PEPE_TOKEN } from '$env/tokens/tokens-erc20/tokens.pepe.env';
-import { BTC_MAINNET_TOKEN, BTC_TESTNET_TOKEN } from '$env/tokens/tokens.btc.env';
+import {BTC_MAINNET_SYMBOL, BTC_MAINNET_TOKEN, BTC_TESTNET_TOKEN} from '$env/tokens/tokens.btc.env';
 import { ETHEREUM_TOKEN } from '$env/tokens/tokens.eth.env';
 import { ICP_TOKEN } from '$env/tokens/tokens.icp.env';
 import { DEPRECATED_SNES } from '$env/tokens/tokens.sns.deprecated.env';
@@ -17,7 +17,7 @@ import type { TokenToggleable } from '$lib/types/token-toggleable';
 import { usdValue } from '$lib/utils/exchange.utils';
 import {
 	filterEnabledTokens,
-	filterTokens,
+	filterTokens, findToken,
 	pinEnabledTokensAtTop,
 	pinTokensWithBalanceAtTop,
 	sortTokens,
@@ -446,5 +446,17 @@ describe('filterTokens', () => {
 		expect(filterTokens({ tokens: [...mockTokens, PEPE_TOKEN], filter: 'ethereum' })).toStrictEqual(
 			[ETHEREUM_TOKEN, PEPE_TOKEN]
 		);
+	});
+});
+
+describe('findToken', () => {
+	it('should return the correct token by symbol', () => {
+		const result = findToken({tokens: mockTokens, symbol: BTC_MAINNET_SYMBOL});
+		expect(result).toEqual(BTC_MAINNET_TOKEN);
+	});
+
+	it('should return undefined if token is not found', () => {
+		const result = findToken({tokens: mockTokens, symbol: 'UNKNOWN_TOKEN'});
+		expect(result).toBeUndefined();
 	});
 });
